Exposure color picker component in geti/ui package
