+++ web_ui/packages/ui/index.ts
@@ -134,6 +134,7 @@ export { type ToggleProps } from '@react-types/checkbox';
 export { Button, ActionButton, type ButtonProps, type ActionButtonProps } from './src/button/button.component';
 export { Checkbox } from './src/checkbox/checkbox.component';
 export { ColorThumb, type ColorThumbProps } from './src/color-thumb/color-thumb.component';
+export { ColorPickerDialog } from './src/color-picker-dialog/color-picker-dialog.component';
 export { DatePicker } from './src/date-picker/date-picker.component';
 export { DateRangePicker } from './src/date-range-picker/date-range-picker.component';
 export { Slider } from './src/slider/slider.component';
