+++ web_ui/packages/ui/src/color-picker-dialog/color-picker-dialog.module.scss
