+++ web_ui/src/pages/create-project/components/project-labels-management/utils.ts
@@ -55,16 +55,3 @@ export const getTutorialCardKey = (
         return TUTORIAL_CARD_KEYS.CREATE_PROJECT_LABELS_CLASSIFICATION_HIERARCHICAL;
     }
 };
-
-export const validateColor = (input: string | undefined) => {
-    const validatedColor = input?.split('#')[1];
-    if (!validatedColor) {
-        return '#000000';
-    } else if (validatedColor.length === 3) {
-        //eslint-disable-next-line max-len
-        return `#${validatedColor[0]}${validatedColor[0]}${validatedColor[1]}${validatedColor[1]}${validatedColor[2]}${validatedColor[2]}`;
-    } else if (/^#[0-9A-F]{6}$/i.test(input)) {
-        return input;
-    }
-    return `#${'0'.repeat(6 - validatedColor.length)}${validatedColor}`;
-};
