+++ web_ui/src/pages/landing-page/landing-page-workspace/components/dataset-import-to-new-project/dataset-import-to-new-project-labels.component.tsx
@@ -3,14 +3,23 @@
 
 import { useCallback, useMemo, useState } from 'react';
 
-import { ActionButton, Checkbox, Divider, Flex, Heading, IllustratedMessage, Text, View } from '@geti/ui';
+import {
+    ActionButton,
+    Checkbox,
+    ColorPickerDialog,
+    Divider,
+    Flex,
+    Heading,
+    IllustratedMessage,
+    Text,
+    View,
+} from '@geti/ui';
 import { ChevronRightSmallLight, LabelGroup, NotFound } from '@geti/ui/icons';
 import { differenceBy, intersectionBy, isEmpty, uniqBy } from 'lodash-es';
 
 import { DATASET_IMPORT_TASK_TYPE } from '../../../../../core/datasets/dataset.enum';
 import { DatasetImportLabel, DatasetImportToNewProjectItem } from '../../../../../core/datasets/dataset.interface';
 import { idMatchingFormat } from '../../../../../test-utils/id-utils';
-import { ColorPickerDialog } from '../../../../create-project/components/project-labels-management/task-labels-management/color-picker-dialog.component';
 import { getLabelFullPath, getLabelsStructureFromPaths, isTaskChainedProject } from './utils';
 
 import classes from './dataset-import-to-new-project.module.scss';
